requirements.txt
absl-py==2.1.0
astunparse==1.6.3
certifi==2025.1.31
charset-normalizer==3.4.1
DateTime==5.5
docutils==0.21.2
flatbuffers==25.2.10
gast==0.6.0
glcontext==3.0.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
keras==3.9.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
moderngl==5.12.0
MouseInfo==0.1.3
namex==0.0.8
noise==1.2.2
numpy==2.1.3
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pillow==11.1.0
protobuf==5.29.3
PyAutoGUI==0.9.54
pygame==2.6.1
PyGetWindow==0.0.9
pyglm==2.8.0
Pygments==2.19.1
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python3-xlib==0.15
pytweening==1.2.0
pytz==2025.1
requests==2.32.3
rich==13.9.4
scipy==1.15.2
six==1.17.0
statistics==1.0.3.5
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
zope.interface==7.2
matplotlib==3.10.1